app/frontend/requirements.txt
streamlit
pillow 